Refactor image writing and counter type in DallE class
Refactored the DallE class. The responsibility of image writing operations was moved from the `DallE` class to the `FileUtils` class. The `getImages` method now returns a long instead of an int. This improves the separation of concerns and allows handling larger amounts of images.

+++ src/test/java/com/kousenit/openai/DallETest.java
@@ -9,7 +9,7 @@ class DallETest {
     @Test
     public void testDallE() {
         DallE dallE = new DallE();
-        int num = dallE.getImages("""
+        long num = dallE.getImages("""
                 A realistic photo of a
                 robot leaping into the air
                 in joy after accomplishing a
